Math type juggling WEIRD SHIT IS HAPPENING WITH OVERFLOW.

+++ openufo.h
@@ -87,7 +87,7 @@ unsigned long flashLEDCurrentState = LOW;
 unsigned long prizeDetectStartMillis = 0;
 
 // Credit Handling
-int totalCredits = 0;					   // Signed so I can detect if the code goofed and went below 0.
+uint16_t totalCredits = 0;
 unsigned long creditDetectStartMillis = 0; // The coin comparitor sends out a 25/50/100ms pulse so we have to time to prevent detecting a long pulse as multiple credits.
 byte creditDetectPulseMs = 100;			   // Set this to match the coin comparitor pulse.
 bool creditSwitchPressed = false;
